// src/common/api/entities/sys/TypeRefs.ts
export type Patch = {
	_type: TypeRef<Patch>;
	_original?: Patch

	_id: Id;
	patchOperation: NumberString;
	attributePath: string;
	value: null | string;
}